split mxu fragment headers
