+++ crates/backend-uzu/src/backends/metal/kernel/gdn/chunked/output_and_state.metal
@@ -3,7 +3,7 @@
 #include "../../common/dsl.h"
 #include "../../common/thread_context.h"
 #include "../../matmul/common/fragment.h"
-#include "../../matmul/common/mxu_fragment_ops.h"
+#include "../../matmul/common/mxu_fragment/ops.h"
 #include "../../matmul/common/simdgroup_fragment_ops.h"
 
 using namespace metal;
