+++ crates/backend-uzu/src/backends/metal/kernel/matmul/gemm/common/mxu_mma_core.h
@@ -5,7 +5,7 @@
 #include "../../../hadamard_transform/hadamard_transform.h"
 #include "../../common/fragment.h"
 #include "gemm_rht.h"
-#include "../../common/mxu_fragment_ops.h"
+#include "../../common/mxu_fragment/ops.h"
 #include "../../common/mxu_gemm_loop.h"
 #include "../../../generated/matmul.h"
 #include "../generated/gemm.h"
